+++ idata-system/src/main/java/cn/zhengcaiyun/idata/system/dal/dao/SysConfigDao.java
@@ -57,7 +57,7 @@ public interface SysConfigDao {
     @SelectProvider(type=SqlProviderAdapter.class, method="select")
     @Results(id="SysConfigResult", value = {
         @Result(column="id", property="id", jdbcType=JdbcType.BIGINT),
-        @Result(column="del", property="del", jdbcType=JdbcType.SMALLINT),
+        @Result(column="del", property="del", jdbcType=JdbcType.TINYINT),
         @Result(column="create_time", property="createTime", jdbcType=JdbcType.TIMESTAMP),
         @Result(column="edit_time", property="editTime", jdbcType=JdbcType.TIMESTAMP),
         @Result(column="key_one", property="keyOne", jdbcType=JdbcType.VARCHAR),
