+++ idata-system/src/main/java/cn/zhengcaiyun/idata/system/dal/model/SysConfig.java
@@ -18,10 +18,10 @@ public class SysConfig {
 
     /**
      * Database Column Remarks:
-     *   是否删除(1:是,其他:否)
+     *   是否删除(1:是,0:否)
      */
     @Generated(value="org.mybatis.generator.api.MyBatisGenerator", comments="Source field: sys_config.del")
-    private Short del;
+    private Integer del;
 
     /**
      * Database Column Remarks:
@@ -62,12 +62,12 @@ public class SysConfig {
     }
 
     @Generated(value="org.mybatis.generator.api.MyBatisGenerator", comments="Source field: sys_config.del")
-    public Short getDel() {
+    public Integer getDel() {
         return del;
     }
 
     @Generated(value="org.mybatis.generator.api.MyBatisGenerator", comments="Source field: sys_config.del")
-    public void setDel(Short del) {
+    public void setDel(Integer del) {
         this.del = del;
     }
