+++ idata-user/src/main/java/cn/zhengcaiyun/idata/user/dal/dao/UacUserTokenDynamicSqlSupport.java
@@ -19,10 +19,10 @@ public final class UacUserTokenDynamicSqlSupport {
 
     /**
      * Database Column Remarks:
-     *   是否删除(1:是,其他:否)
+     *   是否删除(1:是,0:否)
      */
     @Generated(value="org.mybatis.generator.api.MyBatisGenerator", comments="Source field: uac_user_token.del")
-    public static final SqlColumn<Short> del = uacUserToken.del;
+    public static final SqlColumn<Integer> del = uacUserToken.del;
 
     /**
      * Database Column Remarks:
@@ -56,7 +56,7 @@ public final class UacUserTokenDynamicSqlSupport {
     public static final class UacUserToken extends SqlTable {
         public final SqlColumn<Long> id = column("id", JDBCType.BIGINT);
 
-        public final SqlColumn<Short> del = column("del", JDBCType.SMALLINT);
+        public final SqlColumn<Integer> del = column("del", JDBCType.TINYINT);
 
         public final SqlColumn<Date> createTime = column("create_time", JDBCType.TIMESTAMP);
